build: remove FHIR to MR types from the SDK
Ref: metriport/metriport-internal#1182
