+++ packages/api/src/command/medical/patient/consolidated-get.ts
@@ -8,13 +8,7 @@ import {
   ResourceType,
 } from "@medplum/fhirtypes";
 import { ResourceTypeForConsolidation } from "@metriport/api-sdk";
-import { MetriportError } from "@metriport/core/util/error/metriport-error";
-import { makeLambdaClient } from "@metriport/core/external/aws/lambda";
-import {
-  FhirToMedicalRecordPayload,
-  ConsolidationConversionType,
-} from "@metriport/api-sdk/medical/models/fhir";
-import { Config } from "../../../shared/config";
+import { ConsolidationConversionType } from "@metriport/core/domain/fhir";
 import { Patient } from "../../../domain/medical/patient";
 import { QueryProgress } from "../../../domain/medical/query-status";
 import { makeFhirApi } from "../../../external/fhir/api/api-factory";
@@ -26,10 +20,8 @@ import { capture } from "../../../shared/notifications";
 import { emptyFunction, Util } from "../../../shared/util";
 import { updateConsolidatedQueryProgress } from "./append-consolidated-query-progress";
 import { processConsolidatedDataWebhook } from "./consolidated-webhook";
+import { handleBundleToMedicalRecord } from "./convert-fhir-bundle";
 import { getPatientOrFail } from "./get-patient";
-
-const region = Config.getAWSRegion();
-const lambdaClient = makeLambdaClient(region);
 
 export async function startConsolidatedQuery({
   cxId,
@@ -86,6 +78,8 @@ async function getConsolidatedAndSendToCx({
     let bundle = await getConsolidatedPatientData({ patient, resources, dateFrom, dateTo });
 
     if (conversionType) {
+      // If we need to convert to medical record, we also have to update the resulting
+      // FHIR bundle to represent that.
       bundle = await handleBundleToMedicalRecord({
         bundle,
         patient,
@@ -228,115 +222,4 @@ function issueToString(issue: OperationOutcomeIssue): string {
     (issue.diagnostics ? issue.diagnostics.slice(0, 100) + "..." : null) ??
     JSON.stringify(issue)
   );
-}
-
-async function handleBundleToMedicalRecord({
-  bundle,
-  patient,
-  resources,
-  dateFrom,
-  dateTo,
-  conversionType,
-}: {
-  bundle: Bundle<Resource>;
-  patient: Pick<Patient, "id" | "cxId" | "data">;
-  resources?: ResourceTypeForConsolidation[];
-  dateFrom?: string;
-  dateTo?: string;
-  conversionType: ConsolidationConversionType;
-}): Promise<Bundle<Resource>> {
-  const fhir = makeFhirApi(patient.cxId);
-
-  const fhirPatient = await fhir.readResource("Patient", patient.id);
-
-  const bundleWithPatient: Bundle<Resource> = {
-    ...bundle,
-    total: (bundle.total ?? 0) + 1,
-    entry: [
-      {
-        resource: fhirPatient,
-      },
-      ...(bundle.entry ?? []),
-    ],
-  };
-
-  const url = await convertFHIRBundleToMedicalRecord({
-    bundle: bundleWithPatient,
-    patient,
-    resources,
-    dateFrom,
-    dateTo,
-    conversionType,
-  });
-
-  return {
-    resourceType: "Bundle",
-    total: 1,
-    type: "collection",
-    entry: [
-      {
-        resource: {
-          resourceType: "DocumentReference",
-          subject: {
-            reference: `Patient/${patient.id}`,
-          },
-          content: [
-            {
-              attachment: {
-                contentType: `application/${conversionType}`,
-                url: url,
-              },
-            },
-          ],
-        },
-      },
-    ],
-  };
-}
-
-async function convertFHIRBundleToMedicalRecord({
-  bundle,
-  patient,
-  resources,
-  dateFrom,
-  dateTo,
-  conversionType,
-}: {
-  bundle: Bundle<Resource>;
-  patient: Pick<Patient, "id" | "cxId" | "data">;
-  resources?: ResourceTypeForConsolidation[];
-  dateFrom?: string;
-  dateTo?: string;
-  conversionType: ConsolidationConversionType;
-}): Promise<string> {
-  const lambdaName = Config.getFHIRToMedicalRecordLambdaName();
-
-  const payload: FhirToMedicalRecordPayload = {
-    bundle,
-    patientId: patient.id,
-    firstName: patient.data.firstName,
-    cxId: patient.cxId,
-    resources,
-    dateFrom,
-    dateTo,
-    conversionType,
-  };
-
-  const lambdaResult = await lambdaClient
-    .invoke({
-      FunctionName: lambdaName,
-      InvocationType: "RequestResponse",
-      Payload: JSON.stringify(payload),
-    })
-    .promise();
-
-  if (lambdaResult.StatusCode !== 200)
-    throw new MetriportError("Lambda invocation failed", undefined, { lambdaName });
-
-  if (lambdaResult.Payload === undefined)
-    throw new MetriportError("Payload is undefined", undefined, { lambdaName });
-
-  const url = lambdaResult.Payload.toString();
-
-  return url.replace(/['"]+/g, "");
 }
